Version: 0.20.0-beta.2+11 Canonicalize dim_t as the type of every tensor-axis dimension
- Rule, now recorded in CLAUDE.md Code Style: dim_t is the type of anything
  describing a tensor axis -- its extent, a position within it, or a count of
  its elements -- at every API, config, component, and operation-interface
  boundary. size_t never describes a dimension. Token ids are values, not
  extents, and stay out of scope.
- Add narrowToKernelIndex() to Tensor.Types.ixx as the single checked
  dim_t -> int narrowing point, applied at the kernel launch site in CudaGqaOp,
  CudaMhaOp, CudaRopeOp and CudaLpeOp. It throws rather than wrapping: a Gemma 4
  12B embedding table is ~1.0e9 elements against an INT_MAX of 2.1e9. Kernel
  internals and the *.Dispatch / *.Plans layers keep 32-bit indexing unchanged.
- Move the three straggler configs off size_t (Rope, Lpe, TokenEmbedding),
  completing a migration the other sixteen had already made. Required-positive
  validation moves from == 0 to <= 0, and rotary_dim gains the negative guard it
  could not have while unsigned.
- Widen the KV and positional interfaces (IKvCacheLifecycle, IKvInference,
  IPackedKvInference, IPositionalDecode, IPositionalPairedOp), their four
  implementing ops, and RopeCacheRegistry::CacheKey, whose size_t extent fields
  made its brace-init a narrowing error once the config returned dim_t.
- Retire all six REVIEW: markers the drift had accumulated, in
  GroupedQueryAttention, Linear, TokenEmbedding, Rope.Config, GemmaModel and the
  Gemma transformer test. Update the affected tests, whose unsigned literals and
  int casts no longer matched the signed getters.
- Hold Tensor::size() at size_t deliberately, tracked as its own BACKLOG item:
  its blast radius runs through every parameterCount() override and cannot be
  predicted from grep, so it needs a full rebuild as its oracle.

BREAKING: LanguageNetwork::decode, prefillFrom and rewindKvCache now take dim_t;
prefillFrom had already drifted to int64_t while decode stayed int. The
LanguageModel maxSequenceLength and vocabSize virtuals return dim_t. The five
KV/positional operation interfaces take dim_t positions and extents. RopeConfig,
LpeConfig and TokenEmbeddingConfig accessors and fluent setters return and accept
dim_t instead of size_t. xavier() takes dim_t fan_in/fan_out. Overriders and
callers must update in lockstep; every one of these is virtual or brace-init
checked, so the compiler reports any missed site as a hard error.

+++ Mila/Tests/Dnn/Components/Transformers/Gemma/Gemma.Cuda.cpp
@@ -284,7 +284,7 @@ namespace Mila::Tests::Dnn::Components::Transformers::Gemma
         net->synchronize();
 
         // Rewind to the split ([0, kSplit) stays resident) and prefill only the tail.
-        ASSERT_TRUE( net->rewindKvCache( static_cast<int>( kSplit ) ) );
+        ASSERT_TRUE( net->rewindKvCache( kSplit ) );
 
         auto& logits_incremental = net->prefillFrom( tokens, kSplit );
         HostTensor incremental_host( Device::Cpu(), logits_incremental.shape() );
@@ -311,8 +311,8 @@ namespace Mila::Tests::Dnn::Components::Transformers::Gemma
         net->prefill( tokens );
         net->synchronize();
 
-        EXPECT_TRUE( net->rewindKvCache( static_cast<int>( seq_ - 1 ) ) );
-        EXPECT_FALSE( net->rewindKvCache( static_cast<int>( seq_ + 10 ) ) );
+        EXPECT_TRUE( net->rewindKvCache( seq_ - 1 ) );
+        EXPECT_FALSE( net->rewindKvCache( seq_ + 10 ) );
     }
 
     TEST_F( GemmaTransformerCudaTests, PrefillFrom_ThrowsOnOffsetOutsidePrompt )
@@ -334,11 +334,7 @@ namespace Mila::Tests::Dnn::Components::Transformers::Gemma
 
         auto step = makeTokens( batch_, 1 );
 
-        // REVIEW: These static_casts are a smell that the public decode() interface
-        // is not well-aligned with the internal block contract (int vs int64_t).
-        // The public interface should be fixed to take int64_t (dim_t) and avoid these.
-
-        auto& logits = net->decode( step, static_cast<int>( seq_ - 1) );
+        auto& logits = net->decode( step, seq_ - 1 );
 
         EXPECT_EQ( logits.shape(), ( shape_t{ batch_, 1, kVocab } ) );
     }
